--- a/openquake/hazardlib/contexts.py
+++ b/openquake/hazardlib/contexts.py
@@ -579,23 +579,16 @@ class ContextMaker(object):
             setattr(ctx, param, value)
         return ctx
 
-    def get_ctx(self, rup, sites, magdist):
+    def get_ctx(self, rup, sites, distances):
         """
         :returns: a RuptureContext (or None if filtered away)
         """
-        distances = get_distances(rup, sites, 'rrup', self.dcache)
-        mask = distances <= magdist
-        if mask.any():
-            r_sites, distances = sites.filter(mask), distances[mask]
-        else:
-            return
-
         # add distances to the context
         ctx = self.make_rctx(rup)
         ctx.rrup = distances
-        ctx.sites = r_sites
+        ctx.sites = sites
         for param in self.REQUIRES_DISTANCES - {'rrup'}:
-            dists = get_distances(rup, r_sites, param, self.dcache)
+            dists = get_distances(rup, sites, param, self.dcache)
             setattr(ctx, param, dists)
 
         # Equivalent distances
@@ -608,8 +601,8 @@ class ContextMaker(object):
                 ctx.rrup = numpy.sqrt(reqv**2 + rup.hypocenter.depth**2)
 
         # add site parameters
-        for name in r_sites.array.dtype.names:
-            setattr(ctx, name, r_sites[name])
+        for name in sites.array.dtype.names:
+            setattr(ctx, name, sites[name])
 
         # get closest point on the surface
         if len(sites.complete) <= self.max_sites_disagg:
@@ -653,9 +646,13 @@ class ContextMaker(object):
                 if len(rups) == 0:  # may happen in case of min_mag/max_mag
                     continue
                 magdist = self.maximum_distance(rups[0].mag)
-                for rup in rups:
-                    ctx = self.get_ctx(rup, sites, magdist)
-                    if ctx:
+                alldists = [get_distances(rup, sites, 'rrup', self.dcache)
+                            for rup in rups]
+                for rup, dists in zip(rups, alldists):
+                    mask = dists <= magdist
+                    if mask.any():
+                        r_sites, distances = sites.filter(mask), dists[mask]
+                        ctx = self.get_ctx(rup, r_sites, distances)
                         ctx.src_id = src_id
                         ctxs.append(ctx)
         return ctxs