--- a/openquake/hazardlib/geo/surface/planar.py
+++ b/openquake/hazardlib/geo/surface/planar.py
@@ -138,83 +138,92 @@ def _project(self, points):
 
 def get_rrup(planar, xyz):
     """
-    :param planar: a PlanarSurface or a planar array
+    :param planar: a planar array of shape (U, 3)
     :param xyz: an array of euclidean coordinates of shape (N, 3)
-    :returns: N distances for the surface to the points.
+    :returns: (U, N) distances for the surface to the points.
     """
-    width, length = planar.wld[:2]
-    # we project all the points of the mesh on a plane that contains
-    # the surface (translating coordinates of the projections to a local
-    # 2d space) and at the same time calculate the distance to that
-    # plane.
-    dists, xx, yy = _project(planar, xyz)
-    # the actual resulting distance is a square root of squares
-    # of a distance from a point to a plane that contains the surface
-    # and a distance from a projection of that point on that plane
-    # and a surface rectangle. we have former (``dists``), now we need
-    # to find latter.
-    #
-    # we process separately two coordinate components of the point
-    # projection. for abscissa we consider three possible cases:
-    #
-    #  I  . III .  II
-    #     .     .
-    #     0-----+                → x axis direction
-    #     |     |
-    #     +-----+
-    #     .     .
-    #     .     .
-    #
-    mxx = numpy.select(
-        condlist=[
-            # case "I": point on the left hand side from the rectangle
-            xx < 0,
-            # case "II": point is on the right hand side
-            xx > length
-            # default -- case "III": point is in between vertical sides
-        ],
-        choicelist=[
-            # case "I": we need to consider distance between a point
-            # and a line containing left side of the rectangle
-            xx,
-            # case "II": considering a distance between a point and
-            # a line containing the right side
-            xx - length
-        ],
-        # case "III": abscissa doesn't have an effect on a distance
-        # to the rectangle
-        default=0
-    )
-    # for ordinate we do the same operation (again three cases):
-    #
-    #    I
-    #  - - - 0---+ - - -         ↓ y axis direction
-    #   III  |   |
-    #  - - - +---+ - - -
-    #    II
-    #
-    myy = numpy.select(
-        condlist=[
-            # case "I": point is above the rectangle top edge
-            yy < 0,
-            # case "II": point is below the rectangle bottom edge
-            yy > width
-            # default -- case "III": point is in between lines containing
-            # top and bottom edges
-        ],
-        choicelist=[
-            # case "I": considering a distance to a line containing
-            # a top edge
-            yy,
-            # case "II": considering a distance to a line containing
-            # a bottom edge
-            yy - width
-        ],
-        # case "III": ordinate doesn't affect the distance
-        default=0
-    )
-    # combining distance on a plane with distance to a plane
-    return numpy.sqrt(dists ** 2 + mxx ** 2 + myy ** 2)
+    distances = numpy.zeros((len(planar), len(xyz)))
+    for p, point in enumerate(xyz):
+        width = planar.wld[:, 0]  # shape U
+        length = planar.wld[:, 1]  # shape U
+        d = planar.wld[:, 2]  # shape U
+        # we project all the points of the mesh on a plane that contains
+        # the surface (translating coordinates of the projections to a local
+        # 2d space) and at the same time calculate the distance to that
+        # plane.
+        dists = planar.normal @ point + d  # shape U
+        # translate projected points to surface coordinate space, shape (N, 3)
+        mat = point - planar.xyz[:, :, 0]  # shape (U, 3)
+        xx, yy = dot(mat, planar.uv1), dot(mat, planar.uv2)  # shape U each
+
+        # the actual resulting distance is a square root of squares
+        # of a distance from a point to a plane that contains the surface
+        # and a distance from a projection of that point on that plane
+        # and a surface rectangle. we have former (``dists``), now we need
+        # to find latter.
+        #
+        # we process separately two coordinate components of the point
+        # projection. for abscissa we consider three possible cases:
+        #
+        #  I  . III .  II
+        #     .     .
+        #     0-----+                → x axis direction
+        #     |     |
+        #     +-----+
+        #     .     .
+        #     .     .
+        #
+        mxx = numpy.select(
+            condlist=[
+                # case "I": point on the left hand side from the rectangle
+                xx < 0,
+                # case "II": point is on the right hand side
+                xx > length
+                # default -- case "III": point is in between vertical sides
+            ],
+            choicelist=[
+                # case "I": we need to consider distance between a point
+                # and a line containing left side of the rectangle
+                xx,
+                # case "II": considering a distance between a point and
+                # a line containing the right side
+                xx - length
+            ],
+            # case "III": abscissa doesn't have an effect on a distance
+            # to the rectangle
+            default=0
+        )
+        # for ordinate we do the same operation (again three cases):
+        #
+        #    I
+        #  - - - 0---+ - - -         ↓ y axis direction
+        #   III  |   |
+        #  - - - +---+ - - -
+        #    II
+        #
+        myy = numpy.select(
+            condlist=[
+                # case "I": point is above the rectangle top edge
+                yy < 0,
+                # case "II": point is below the rectangle bottom edge
+                yy > width
+                # default -- case "III": point is in between lines containing
+                # top and bottom edges
+            ],
+            choicelist=[
+                # case "I": considering a distance to a line containing
+                # a top edge
+                yy,
+                # case "II": considering a distance to a line containing
+                # a bottom edge
+                yy - width
+            ],
+            # case "III": ordinate doesn't affect the distance
+            default=0
+        )
+        # combining distance on a plane with distance to a plane
+        distances[:, p] = numpy.sqrt(dists ** 2 + mxx ** 2 + myy ** 2)
+    return distances
 
 
 class PlanarSurface(BaseSurface):
@@ -383,6 +392,7 @@ class PlanarSurface(BaseSurface):
         self = object.__new__(PlanarSurface)
         self.strike = strike
         self.dip = dip
+        self.array = planar_array
         for par in planar_array.dtype.names:
             setattr(self, par, planar_array[par])
         return self
@@ -423,9 +433,9 @@ class PlanarSurface(BaseSurface):
         Prepare everything needed for projecting arbitrary points on a plane
         containing the surface.
         """
-        planar_array = build_planar_array(self.corners, check=check)
-        for par in planar_array.dtype.names:
-            setattr(self, par, planar_array[par])
+        self.array = build_planar_array(self.corners, check=check)
+        for par in self.array.dtype.names:
+            setattr(self, par, self.array[par])
 
     def translate(self, p1, p2):
         """
@@ -518,7 +528,7 @@ class PlanarSurface(BaseSurface):
         This is an optimized version specific to planar surface that doesn't
         make use of the mesh.
         """
-        return get_rrup(self, mesh.xyz)
+        return get_rrup(self.array.reshape(1, 3), mesh.xyz)
 
     def get_closest_points(self, mesh):
         """