… openquake/commonlib/tests/source_test.py