Part of #7745. Used numba to speedup the computation of the rrup distances. For the test share_small on my workstation I get
```
# before
| calc_56821, maxmem=6.9 GB  | time_sec | memory_mb | counts |
|----------------------------+----------+-----------+--------|
| total classical            | 5_100    | 308.1     | 251    |
| make_contexts              | 4_569    | 527.8     | 18_865 |
| iter_ruptures              | 1_487    | 0.0       | 18_865 |
| ClassicalCalculator.run    | 632.2    | 1_136     | 1      |

# after
| calc_56822, maxmem=7.0 GB  | time_sec | memory_mb | counts  |
|----------------------------+----------+-----------+---------|
| total classical            | 4_315    | 321.2     | 251     |
| make_contexts              | 3_784    | 576.3     | 18_865  |
| iter_ruptures              | 1_498    | 0.0       | 18_865  |
| ClassicalCalculator.run    | 543.6    | 1_130     | 1       |
```